Convert to GitHub Flavored Markdown
Use fenced code blocks and unordered lists rather than definition lists.
Also update headers so the title is an h1 and all others are h2 or
smaller.

While at it, update the description of semver to comply with semver v2
and lowercase distribution names, which as been the norm since at least
2013.

Otherwise clean up a few minor typos and formatting errors, and convert
all links to reference links.