Implement JSON Schema for PGXN Meta Spec v1
Based on the descriptions in `spec.md` and the validation in
pgxn/pgxn-meta-validator, write JSON Schema v2020 schemas to validate
PGXN Meta Spec `META.json` files. The implementation should be at least
as good as the old one, now portable for use in other languages.

In the process, correct a few issues and errors discovered in `spec.md`
and increment its version to v1.0.2.

Add a test suite in Rust that uses the boon crate to verify that the
schemas, their IDs, and their examples are valid. This sets the stage
for writing a validation utility in Rust.

To keep things tidy and somewhat standardized, add a pre-commit
configuration to lint the JSON and Rust code.

Then setup a GitHub workflow to run the tests on macOS, Windows, and
Linux, and the pre-commit linting on Linux. Overkill for now, since
there is no Rust library, just tests for the JSON Schema, but likely to
be needed before long.