Implement v2 Spec JSON Schema
Based on pgxn/rfcs#3. As with the v1 schemas, use JSON Schema v2020 for
the v2 schemas.

Fix a few minor issues with the v1 schema, as well. Add full test suite
in Rust; move code shared between the v1 and v2 tests to
`tests/common/mod.rs`. This includes custom format functions not
required for v1, including spdx version expression validation with the
`spdx` crate, and path validation to ensure no `..` is included in
paths.