Turn schema tests into unit tests
It was a bit fussy to have them as integration tests, and that's not
really what they are anyway. So move them into `src/tests` and otherwise
rejigger things only enough for them to have access to the private stuff
they need.

Also move the test corpus into the top-level `corpus` directory.

While at it, move the compiler module under valid, as it's only needed
by the valid module and the unit tests. Remove it from the root module
and instead load it publicly only for tests.