Add release workflows
Follow the example of houseabsolute/ubi to use cross-rs/cross to
cross-compile binaries for multiple platforms and architectures. Use
`musl` where available to support both `glibc` and `musl` Linux
distributions. Test for every push against stable, beta, and nightly,
and build archives that contain the license, readme, and change log, as
well as the `pgxn_meta` binary.

Also add a workflow to publish on crates.io, and add a dependency check
step to the lint and test workflow.

Add `CHANGELOG.md` for tracking changes, and update `Cargo.toml` for
distribution on crates.io.

Revise the README for the crate and binary.