Test and fix serialization
Add `TryFrom<Meta> for Value` and `TryFrom<&Meta> for String` to handle
serialization.

Then add tests to ensure that Meta object loaded from v2 meta files
deserialize into the same data structure. Done mainly by omitting
Optional fields set to `None`, but also by eliminating the
`meta-spec.url` default value and a custom `x_` property, which will be
restored in a future commit.

In the process, add support for badges in the resources object. This was
an oversight in e803409.

While at it, change a parameter instance of `&Vec<T>` to `&[T]`.