Add TryFrom multiple JSON values
Assume the first value is the primary metadata, generally included in a
distribution. Merge subsequent values into that first value via the [RFC
7396] merge pattern. This is ideal for changing objects, but arrays must
be replaced in their entirety. When all patches have been applied, then
validate the final document and deserialize it into a Meta object.Since

In order to convert v1 metadata to a JSON value, refactor
`src/meta/v1/mod.rs` by adding `to_v2()`, which returns a JSON Value,
and use it in the existing `from_value()` function. While at it, use a
reference to the v1 metadata in `to_v2()`, since its data is copied to
the v2 result.

Use the json_merge crate to do the merging. The [implementation] is dead
simple, so it's kind of overkill, but it's nice to have the crate on
hand in cas we later decide to support the [JSON Patch] format, as well.

The new trait implementation, `impl TryFrom<&[&Value]> for Meta`, needs
to fetch the meta spec version from the first Value passed to it. So
extract that pattern from `src/valid/mod.rs` and move it to a new `util`
module that's private to the crate. Add a slew of tests for its
behavior, as well.

Add tests to ensure full coverage and proper patch behavior for v1 and
v2 patches, as well as a real-world fairly complicated example patching
the pg_partman v2 metadata to the anticipated v2 metadata.

Since the util function, `get_version`, only ever returns 1 or 2, in the
branching on its value panic with `unreachable!()` for any other value.
This should only happen in some far off future where a v3 is added. To
keep test coverage at 100%, tweak the coverage reporting to ignore lines
that use `unreachable!()`, which are untestable.

  [implementation]: https://github.com/idubrov/json-patch/blob/v2.0.0/src/lib.rs#L660-L677
  [JSON Patch]: https://docs.rs/json-patch/latest/json_patch/fn.patch.html