Add read-only accessor methods to public objects
The intention is that these are read-only objects, so don't expose the
fields, which are validated on loading from JSON, because one could
create mutable instances and modify them without validation.

So add accessor methods to all of the fields in the structures that make
up distribution metadata. Each returns a borrowed reference to the data
in the field, useful for the lifetime of the object. Since we expect
these objects to be loaded early on to be used for the lifetime of
working with the distributions they represent, this should be more than
adequate.

While at it:

*   Return a semver::Version for the meta-spec version field.
*   Make the enums public and add `Display` implementations for each.
*   Don't store Variation Dependencies in a Box, there is no need.
*   Add introductory documentation.
*   Add tests for all of the new accessors.