Add v1-v2 release conversion
Dupe the pattern from the `meta` package by adding `v1` and `v2`
sub-packages that know how to parse into the canonical form. Make
`meta::v1` public to the crate so that `release::v1` can use it to
migrate everything but the release metadata, then add `v1_to_v2_release`
to handle that bit.

The v2 `release` object takes the form of a JWS-JS data structure, but
since we're not signing anything yet, and this is just to get things to
load properly, for now just generate random values for the JWS header
and signature.

Enable the v1 corpus tests and add a test for merging v1 release
metadata.

While at it, change the `uri` property to start with `dist` instead of
`/dist`, because the latter implies relative to the root of a domain,
and mirrors may not serve from a domain root.