Add certs to release package
Add the `certs` field to the `Release` struct and modify the `release`
field to contain `ReleasePayload` parsed from the `pgxn` field of the
`certs` hash map. Since the `release` field does not exist in the JSON,
add custom deserialization to the `Release` struct that parses only the
`dist` and `certs` fields and then extracts the `pgxn` JWS data and
validates its base64-encoded JSON payload. No JWS validation yet, as the
key management has not yet been worked out.

Revise the `v1_to_v2_release` method to build an appropriate Base
64-encoded payload and mock signature from v1 release metadata. Add
tests to ensure that the keys are included in the JSON in unicode code
point order (alphabetical, since it's all lowercase ASCII) and that
there are no spaces.

Update the release tests for the new properties as well as the `certs`
JSON schema.