Add error module and wrap all valid module errors
Add the `error` module, defining an Error enum using `thiserror`. Create
variants for all of the errors generated by the `valid` module, and
replace its custom `ValidationError` enum. Includes tests for each
variant. Then teach that module to return only our errors.

Note the special treatment for these error conversions:

*   `boon::ValidationError`s use lifetimes to hang on to values from the
    boon compiler and from the JSON that fails validation. This makes it
    quite tricky to use separate from those values. But since, so far,
    we just want to print JSON Schema validation errors, add a `From`
    trait implementation that simply copies its string value into our
    error type. This allows all the lifetime annotations to be removed
    from the `valid` crate.

*   `wax::GlobError` is an enum that wraps the two types that the `wax`
    create generates. To coerce those two types, add a `From` trait for
    each.

While at it, improve the error generation for the custom `license`
format so that the error message generated by boon is more legible and
useful. Previously it just showed the value that failed --- twice ---
but now it also shows the "reason" for the error.

Also: change the behavior of `is_license` and `is_path` to return
`Ok(())` if the value is not a string, following the example of the
format validation functions in boon itself. New tests for the `license`
and `path` custom formats validate the usability of the error messages.