Extend release validation to support v1 API files
The API Server adds additional properties to the [Meta API], an which
here is called Release (because it contains release metadata). This
caused validation to fail for metadata fetched from api.pgxn.org.

Extend the validation of Release metadata to optionally include these
properties. Define a new v1 schema, `api-extras.schema.json`, and add it
to `release.schema.json`. Like the existing schemas, it applies strict
validation for the new properties.

Since the new schemas validate file paths, copy `path.schema.json` from
the v2 schemas and apply it to the other paths in the v1 schema.

Increment to v0.6.0 and prep for release.

  [Meta API]: https://github.com/pgxn/pgxn-api/wiki/meta-api#api-server-structure