Add tag limit and glob format
Use [wax] to validate globs. Since it checks for both `.` and `..`
components, disallow both for paths and globs, except at the start of an
expression.  May or may not switch away from wax in the future, since
its syntax is somewhat different from gitignore. For now, though, the
RFC has been [updated] to reflect this change.

Also limit the number of tags to 32, also reflecting a recent RFC
change.

  [wax]: https://crates.io/crates/wax
  [updated]: https://github.com/pgxn/rfcs/pull/3/commits/4021dee