Improve test organization and thus coverage
Move all tests to separate `tests` modules. Most were already, but the
tests for `main`, `valid`, and `valid/compiler` were not. This
organization simplifies the `grcov` options. Alas, `src/tests.rs` by
default resolves to the same thing as `src/tests/mod.rs` (the latter
being integration tests), so add `#[path]` macros to point to the proper
files for each.

Remove the doc comment from `From<boon::CompileError> for Error` as it
somehow mucked with line numbering. With it gone, the line numbers are
correct and the proper coverage report results.

Only one line remains that's considered uncovered, the `#[derive]` for
the `Digests` struct in `release/mod.rs`. It thinks that `Deserialize`
is uncovered, even though the tests deserialize to `Digests` repeatedly.
Left a comment there.